Extract PropertiesTemplateLookupSheetExport for PSR-4 autoload.
Fixes Class not found when regular PropertiesExport builds the city/district lookup sheet.
